Move the code for the TestMeterProvider to
separate file
